First revision, huge update, project is now working at its minimum
Updated the readme to a template telling more information about the
project, bonus information, to-do list and credits.
Updated the index.php to show the wanted content together with the two
css files in black and white and in colors.
Choose to use srobbin's backstretch javascript to form the background.
And made functions.js to control the switching between black'n'white
and color.
Took some photos of myself to form the background and made some icons
to form the mail and social buttons. Placed it in the images folder.
Made CONFIG.php to host all the settings of the personal page, and made
it so this is the only file you have to edit (+ 2 new background
images), to make the website complete.
Made helper.php to support the complex tasks of the configuration.